# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IDP Perimeter API
  version: 0.3.0
  description: >
    Контракт периметра (портал ↔ API-шлюз) платформы IDP. Источник правды для
    кодогена TS-клиента и zod-схем.
servers:
  - url: /api
paths:
...
  /projects/{project}/services:
    get:
      operationId: listServices
      summary: Список сервисов проекта с keyset-пагинацией
      description: >
        Возвращает страницу сервисов проекта с keyset-пагинацией
        (page_size/page_token). Требует право list на project:{project}; отказ
        RBAC → 403. Курсор непрозрачен и пробрасывается без интерпретации.
      parameters:
        - $ref: "#/components/parameters/ProjectPath"
        - name: page_size
          in: query
          required: false
          description: >
            Желаемый размер страницы; сервер клампит к допустимому пределу.
            Значение должно укладываться в int32 (1..2147483647); выход за
            диапазон — 400.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 2147483647
        - name: page_token
          in: query
          required: false
          description: >
            Непрозрачный курсор продолжения из предыдущего ответа; пустой —
            с начала выборки.
          schema:
            type: string
      responses:
        "200":
          description: Страница сервисов проекта
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createService
      summary: Запуск создания сервиса в проекте
      description: >
        Запускает создание сервиса в проекте (асинхронная провизия репозитория,
        образов и секретов). Запись каталога фиксируется со статусом creating;
        требует право create на project:{project} (отказ RBAC → 403). Занятое имя
        в проекте → 409.
      parameters:
        - $ref: "#/components/parameters/ProjectPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Создание запущено, запись зафиксирована со статусом creating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateServiceResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/services/{name}:
    get:
      operationId: getService
      summary: Чтение статуса одного сервиса проекта
      description: >
        Возвращает текущее состояние сервиса (статус, владельцы, версия). Требует
        право read на project:{project} (отказ RBAC → 403); отсутствие сервиса →
        404.
      parameters:
        - $ref: "#/components/parameters/ProjectPath"
        - $ref: "#/components/parameters/ServiceNamePath"
      responses:
        "200":
          description: Текущее состояние сервиса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceSummary"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{project}/services/{name}/decommission:
    post:
      operationId: decommissionService
      summary: Вывод сервиса из эксплуатации (soft delete)
      description: >
        Выводит активный сервис из эксплуатации (soft delete): данные каталога
        сохраняются, статус переводится в decommissioned, во внешних системах
        отзывается доступ (GitLab archive, Harbor read-only, Vault revoke).
        Тело несёт явное предусловие load_drained (нагрузка снята из K8s, ADR-0012).
        Идемпотентно: повтор на уже выведенном сервисе → 200. Допустим только
        исходный статус active; предусловие не выполнено (нагрузка не снята или
        статус creating/failed) → 422; конкурентная смена статуса → 409; отсутствие
        сервиса → 404; отказ RBAC (право decommission) → 403.
      parameters:
        - $ref: "#/components/parameters/ProjectPath"
        - $ref: "#/components/parameters/ServiceNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecommissionServiceRequest"
      responses:
        "200":
          description: Вывод из эксплуатации запущен (или сервис уже выведен)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PreconditionFailed"
  /projects/{project}/services/{name}/owners:
    put:
      operationId: setServiceOwners
      summary: Декларативная замена набора владельцев сервиса
      description: >
        Заменяет набор владельцев целиком (идемпотентно). Клиент передаёт полный
        желаемый набор owners и текущую версию owners_version
        (optimistic-concurrency). Несовпадение версии → 409; отсутствие сервиса →
        404; отказ RBAC (право change_owners) → 403.
      parameters:
        - $ref: "#/components/parameters/ProjectPath"
        - $ref: "#/components/parameters/ServiceNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetServiceOwnersRequest"
      responses:
        "200":
          description: Изменение владельцев запущено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetServiceOwnersResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{project}/services/{name}/transfer:
    post:
      operationId: transferService
      summary: Перенос сервиса в другой проект
      description: >
        Переносит сервис в другой проект (смена project-владельца): id записи
        каталога сохраняется, меняется проект (source→target), владельцы переезжают
        вместе с записью. Перенос выполняется асинхронно (Saga) с ТОЧКОЙ НЕВОЗВРАТА
        на переносе репозитория GitLab: каталог active→transferring → transfer
        GitLab → миграция путей Vault → метаданные Harbor → каталог
        transferring→active (project=target) → перенос ролей владельцев в IDM.
        Требует ДВУХ прав: transfer на исходном проекте И transfer_in на целевом
        (ADR-0013). Идемпотентно: повтор на уже перенесённом сервисе → 200.
        Допустим только исходный статус active; недопустимый статус → 422; занятое
        имя в target или конкурентная смена → 409; отсутствие сервиса → 404; отказ
        RBAC (transfer/transfer_in) → 403.
      parameters:
        - $ref: "#/components/parameters/ProjectPath"
        - $ref: "#/components/parameters/ServiceNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferServiceRequest"
      responses:
        "200":
          description: Перенос запущен (или сервис уже перенесён)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PreconditionFailed"
...
components:
  parameters:
    ProjectPath:
      name: project
      in: path
      required: true
      description: Идентификатор проекта-владельца.
      schema:
        type: string
        minLength: 1
    ServiceNamePath:
      name: name
      in: path
      required: true
      description: Имя сервиса внутри проекта.
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Некорректный запрос (валидация входных данных)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Конфликт состояния (например, имя сервиса уже занято)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Доступ запрещён (RBAC, fail-closed)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PreconditionFailed:
      description: >
        Предусловие операции не выполнено (например, нагрузка не снята из K8s или
        недопустимый исходный статус сервиса).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ServiceStatus:
      type: string
      description: Статус сервиса в каталоге (ADR-0004).
      enum: [creating, active, decommissioned, failed, transferring]
    ServiceSummary:
      type: object
      required: [project, name, status, owners, owners_version]
      properties:
        project:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/ServiceStatus"
        owners:
          type: array
          description: Набор владельцев сервиса (детерминированный порядок).
          items:
            type: string
        owners_version:
          type: integer
          format: int64
          description: Версия набора владельцев для optimistic-concurrency.
        decommissioned_at:
          type: string
          description: >
            Момент вывода сервиса из эксплуатации (RFC3339); отсутствует/пусто,
            если сервис не выведен (ADR-0012).
    ServiceList:
      type: object
      required: [services, next_page_token]
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceSummary"
        next_page_token:
          type: string
          description: Курсор следующей страницы; пустой — конец выборки.
    CreateServiceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          description: Имя создаваемого сервиса внутри проекта.
    CreateServiceResult:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          description: Идентификатор созданной записи каталога.
        status:
          $ref: "#/components/schemas/ServiceStatus"
    SetServiceOwnersRequest:
      type: object
      required: [owners, owners_version]
      properties:
        owners:
          type: array
          description: >
            Полный желаемый набор владельцев (декларативно). Без пустых строк и
            дублей; сервер вычисляет diff против текущего состояния.
          items:
            type: string
            minLength: 1
        owners_version:
          type: integer
          format: int64
          description: Ожидаемая версия набора владельцев (optimistic-concurrency).
    SetServiceOwnersResult:
      type: object
      required: [owners, owners_version]
      properties:
        owners:
          type: array
          description: Итоговый набор владельцев после применения.
          items:
            type: string
        owners_version:
          type: integer
          format: int64
          description: Новая версия набора владельцев.
    DecommissionServiceRequest:
      type: object
      required: [load_drained]
      properties:
        load_drained:
          type: boolean
          description: >
            Явное предусловие: нагрузка сервиса снята из K8s. В MVP K8s-worker
            отсутствует, поэтому снятие подтверждается клиентом (ADR-0012).
            false → 422 до любых побочных эффектов.
    TransferServiceRequest:
      type: object
      required: [target_project]
      properties:
        target_project:
          type: string
          minLength: 1
          description: >
            Целевой проект-владелец (target), в который переносится сервис. Должен
            отличаться от исходного проекта; пара (target_project, name) должна
            быть свободна, иначе 409.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: >
            Стабильное сообщение об ошибке для клиента. Внутренние детали
            (текст gRPC-ошибки) сюда НЕ попадают — только в лог сервера.